Release nxtvepg-2.8.0 (2007-Dec-31)
Changes between 2.7.7 and 2.8.0 (as per file CHANGES):
- possibility to compile separate daemon "nxtvepgd" without GUI (mainly
  intended for Linux-based media appliances without X11 libraries)
- import and merge of EPG data from XMLTV files
- grabbing EPG data from regular teletext pages (still experimental and
  hence disabled by default)
- Windows: extended the communication protocol for TV applications to
  allow to transport more EPG data (more details, multiple programmes)
  and to use the nxtvepg daemon; requires new plugin for K!TV
  (included in win32 package)
- UNIX: Implemented new communication protocol for TV applications which
  allows to more EPG data (Download the tvsim package to get a demo) Not
  yet used by any TV applications.
- UNIX: added possibility to use an external program to display EPG info
  popups to TV app. interaction configuration
- added new command line option -provscan de|fr|at|ch: runs an EPG
  scan without GUI. It can also be used by nxtvepgd, so nxtvepg can
  now be set up without ever using the GUI (except for some advanced
  acquisition options)
- added command line option -remctrl for UNIX (was supported on WIN32 only)
- Added support for export in HTML format via the command line: -dump html
  (does not support all features available in the GUI though.)
- WIN32: added support for external WDM driver library (note: DLL is not
  included, see download page; still experimental)
- Upgrade of Tcl/Tk library from version 8.3.4 to 8.5.0 (to benefit from
  this UNIX users must also upgrade their installed Tcl/Tk libraries,
  which is highly recommended due to improvements in widget appearences)

The new features in this release were developed for 3 years (2005-2007),
in parallel to main lines releases 2.7.4 - 2.7.7.

See file CHANGES.RCS for a detailed log of revisions of all files.

+++ epgdb/epgnetio.c
@@ -22,7 +22,7 @@
  *  Author:
  *          Tom Zoerner
  *
- *  $Id: epgnetio.c,v 1.37 2004/06/19 19:55:28 tom Exp tom $
+ *  $Id: epgnetio.c,v 1.38 2007/12/29 22:28:07 tom Exp tom $
  */
 
 #define DEBUG_SWITCH DEBUG_SWITCH_EPGDB
@@ -551,7 +551,7 @@ void EpgNetIo_WriteMsg( EPGNETIO_STATE * pIO, EPGNETIO_MSG_TYPE type, uint msgLe
    assert((pIO->writeLen == 0) && (pIO->pWriteBuf == NULL));
    assert((msgLen == 0) || (pMsg != NULL));
 
-   dprintf2("EpgNetIo-WriteMsg: msg type %d, len %d\n", type, sizeof(EPGNETIO_MSG_HEADER) + msgLen);
+   dprintf2("EpgNetIo-WriteMsg: msg type %d, len %d\n", type, (int)sizeof(EPGNETIO_MSG_HEADER) + msgLen);
 
    pIO->pWriteBuf    = pMsg;
    pIO->freeWriteBuf = freeBuf;
@@ -576,7 +576,7 @@ bool EpgNetIo_WriteEpgQueue( EPGNETIO_STATE * pIO, EPGDB_QUEUE * pOutQueue )
 
    while ((pBlock = EpgDbQueue_Get(pOutQueue)) != NULL)
    {
-      dprintf2("EpgNetIo-WriteEpgQueue: fd %d: write EPG block, msg size %d\n", pIO->sock_fd, sizeof(EPGNETIO_MSG_HEADER) + pBlock->size + BLK_UNION_OFF);
+      dprintf2("EpgNetIo-WriteEpgQueue: fd %d: write EPG block, msg size %d\n", pIO->sock_fd, (int)sizeof(EPGNETIO_MSG_HEADER) + pBlock->size + BLK_UNION_OFF);
 
       pIO->pWriteBuf        = (void *) pBlock;
       pIO->freeWriteBuf     = TRUE;
@@ -656,7 +656,7 @@ bool EpgNetIo_WriteTscQueue( EPGNETIO_STATE * pIO, EPGDB_PI_TSC * pTscQueue )
 
    while ((pTscBuf = EpgTscQueue_PopBuffer(pTscQueue, &bufLen)) != NULL)
    {
-      dprintf2("EpgNetIo-WriteTscQueue: fd %d: write timescale buffer, msg size %d\n", pIO->sock_fd, sizeof(EPGNETIO_MSG_HEADER) + bufLen);
+      dprintf2("EpgNetIo-WriteTscQueue: fd %d: write timescale buffer, msg size %d\n", pIO->sock_fd, (int)sizeof(EPGNETIO_MSG_HEADER) + bufLen);
 
       pIO->pWriteBuf    = (void *) pTscBuf;
       pIO->freeWriteBuf = TRUE;
